backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
pydantic-settings==2.4.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-huggingface
langchain-community
sentence-transformers
duckdb
python-dotenv==1.0.1
tiktoken>=0.7.0
python-multipart==0.0.9
PyPDF2==3.0.1
docx2txt==0.8
python-pptx==0.6.23
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.27.0
pytest==8.0.0

